mooved command fief to town
setting a food stock is relating to a Town
Correcting also the command function, to set the fief object properly, it is an abstract class of Settlement Component, and we should pass by the Town
